@@ -27,19 +27,20 @@ dependencies:
     sdk: flutter
   flutter_spinkit: "^4.1.2"
   http: ^0.12.1
-  provider: ^4.3.1
+  provider: ^5.0.0
   flutter_login: ^1.0.14
   carousel_slider: ^2.2.1
   json_annotation: ^3.0.1
-  flutter_form_builder: ^3.11.5
-  shared_preferences: ^0.5.8
+  flutter_form_builder: ^6.0.1
+  shared_preferences: ^2.0.6
   flutter_verification_code: ^0.2.0
-  google_maps_flutter: ^0.5.28+1
+  google_maps_flutter: ^2.0.6
   geolocator: ^5.3.2+2
   package_info: ^0.4.1
   share: '>=0.6.4+3 <2.0.0'
-  emojis: ^0.9.3
+  # emojis: ^0.9.3
   url_launcher: ^5.7.10
+  country_pickers: ^2.0.0
   
 
 
